sol_v1/requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv
aiohttp>=3.8.0
aiosqlite>=0.19.0
websockets>=11.0
fastapi>=0.110.0
uvicorn>=0.27.0
jinja2>=3.1.0
python-multipart>=0.0.9
